+++ plugins/json.js
@@ -2,7 +2,7 @@ import json from "../loaders/json.js";
 
 export default function () {
   return (site) => {
-    site.data([".json"], json);
-    site.load([".tmpl.json"], json);
+    site.loadData([".json"], json);
+    site.loadPages([".tmpl.json"], json);
   };
 }