+++ plugins/yaml.js
@@ -2,7 +2,7 @@ import yaml from "../loaders/yaml.js";
 
 export default function () {
   return (site) => {
-    site.load([".yml", ".yaml"], yaml);
-    site.data([".yml", ".yaml"], yaml);
+    site.loadData([".yml", ".yaml"], yaml);
+    site.loadPages([".yml", ".yaml"], yaml);
   };
 }