+++ CHANGELOG.md
@@ -6,7 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.8.0] - Unreleased
+### Added
+- New `loadAssets()` to register assets loaders
+
 ### Changed
+- Renamed `load()` to `loadPages()` and removed the `asset` argument.
+- Renamed `data()` to `loadData()`
 - Renamed `helper()` to `data()`
 
 ## [0.7.3] - 2020-10-17
@@ -134,6 +139,7 @@ First version
 
 [#1]: https://github.com/oscarotero/lume/issues/1
 
+[0.8.0]: https://github.com/oscarotero/lume/compare/v0.7.3...HEAD
 [0.7.3]: https://github.com/oscarotero/lume/compare/v0.7.2...v0.7.3
 [0.7.2]: https://github.com/oscarotero/lume/compare/v0.7.1...v0.7.2
 [0.7.1]: https://github.com/oscarotero/lume/compare/v0.7.0...v0.7.1